requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
automationhat==3.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
